Fix missing deallocation for RegOpenKeyExW
