+++ src/windows/system.rs
@@ -12,26 +12,22 @@ use crate::sys::disk::{get_disks, Disk};
 use crate::sys::process::{get_start_time, update_memory, Process};
 use crate::sys::tools::*;
 use crate::sys::users::get_users;
-use crate::sys::utils::get_now;
+use crate::sys::utils::{get_now, get_reg_string_value, get_reg_value_u32};
 
 use crate::utils::into_iter;
 
 use std::cell::UnsafeCell;
 use std::collections::HashMap;
-use std::ffi::OsStr;
 use std::mem::{size_of, zeroed};
-use std::os::windows::ffi::OsStrExt;
-use std::slice::from_raw_parts;
 use std::time::{Duration, SystemTime};
 
 use ntapi::ntexapi::{
     NtQuerySystemInformation, SystemProcessInformation, SYSTEM_PROCESS_INFORMATION,
 };
 use winapi::ctypes::wchar_t;
-use winapi::shared::minwindef::{DWORD, FALSE, HKEY, LPBYTE, TRUE};
+use winapi::shared::minwindef::{FALSE, TRUE};
 use winapi::shared::ntdef::{PVOID, ULONG};
 use winapi::shared::ntstatus::STATUS_INFO_LENGTH_MISMATCH;
-use winapi::shared::winerror;
 use winapi::um::minwinbase::STILL_ACTIVE;
 use winapi::um::processthreadsapi::GetExitCodeProcess;
 use winapi::um::psapi::{GetPerformanceInfo, PERFORMANCE_INFORMATION};
@@ -39,8 +35,7 @@ use winapi::um::sysinfoapi::{
     ComputerNamePhysicalDnsHostname, GetComputerNameExW, GetTickCount64, GlobalMemoryStatusEx,
     MEMORYSTATUSEX,
 };
-use winapi::um::winnt::{HANDLE, KEY_READ};
-use winapi::um::winreg::{RegOpenKeyExW, RegQueryValueExW};
+use winapi::um::winnt::HANDLE;
 
 declare_signals! {
     (),
@@ -562,84 +557,6 @@ pub(crate) fn get_process_name(process: &SYSTEM_PROCESS_INFORMATION, process_id:
             );
 
             String::from_utf16_lossy(slice)
-        }
-    }
-}
-
-fn utf16_str<S: AsRef<OsStr> + ?Sized>(text: &S) -> Vec<u16> {
-    OsStr::new(text)
-        .encode_wide()
-        .chain(Some(0).into_iter())
-        .collect::<Vec<_>>()
-}
-
-fn get_reg_string_value(hkey: HKEY, path: &str, field_name: &str) -> Option<String> {
-    let c_path = utf16_str(path);
-    let c_field_name = utf16_str(field_name);
-
-    let mut new_hkey: HKEY = std::ptr::null_mut();
-    unsafe {
-        if RegOpenKeyExW(hkey, c_path.as_ptr(), 0, KEY_READ, &mut new_hkey) != 0 {
-            return None;
-        }
-
-        let mut buf_len: DWORD = 2048;
-        let mut buf_type: DWORD = 0;
-        let mut buf: Vec<u8> = Vec::with_capacity(buf_len as usize);
-        loop {
-            match RegQueryValueExW(
-                new_hkey,
-                c_field_name.as_ptr(),
-                std::ptr::null_mut(),
-                &mut buf_type,
-                buf.as_mut_ptr() as LPBYTE,
-                &mut buf_len,
-            ) as DWORD
-            {
-                0 => break,
-                winerror::ERROR_MORE_DATA => {
-                    buf.reserve(buf_len as _);
-                }
-                _ => return None,
-            }
-        }
-
-        buf.set_len(buf_len as _);
-
-        let words = from_raw_parts(buf.as_ptr() as *const u16, buf.len() / 2);
-        let mut s = String::from_utf16_lossy(words);
-        while s.ends_with('\u{0}') {
-            s.pop();
-        }
-        Some(s)
-    }
-}
-
-fn get_reg_value_u32(hkey: HKEY, path: &str, field_name: &str) -> Option<[u8; 4]> {
-    let c_path = utf16_str(path);
-    let c_field_name = utf16_str(field_name);
-
-    let mut new_hkey: HKEY = std::ptr::null_mut();
-    unsafe {
-        if RegOpenKeyExW(hkey, c_path.as_ptr(), 0, KEY_READ, &mut new_hkey) != 0 {
-            return None;
-        }
-
-        let mut buf_len: DWORD = 4;
-        let mut buf_type: DWORD = 0;
-        let mut buf = [0u8; 4];
-
-        match RegQueryValueExW(
-            new_hkey,
-            c_field_name.as_ptr(),
-            std::ptr::null_mut(),
-            &mut buf_type,
-            buf.as_mut_ptr() as LPBYTE,
-            &mut buf_len,
-        ) as DWORD
-        {
-            0 => Some(buf),
-            _ => None,
         }
     }
 }
